+++ src/Aks/Aks/Commands/CreateOrUpdateKubeBase.cs
@@ -48,7 +48,6 @@ namespace Microsoft.Azure.Commands.Aks
     {
         protected const string DefaultParamSet = "defaultParameterSet";
         protected readonly Regex DnsRegex = new Regex("[^A-Za-z0-9-]");
-        private const string SUPPRESS_ERROR_OR_WARNING_MESSAGE_ENV_VARIABLE_NAME = "SuppressAzurePowerShellBreakingChangeWarnings";
 
         [Parameter(
             Position = 0,
@@ -264,22 +263,7 @@ namespace Microsoft.Azure.Commands.Aks
                 if (clientSecret == null)
                 {
                     clientSecret = RandomBase64String(16);
-                }
-
-                bool supressWarningOrError = false;
-
-                try
-                {
-                    supressWarningOrError = bool.Parse(Environment.GetEnvironmentVariable(SUPPRESS_ERROR_OR_WARNING_MESSAGE_ENV_VARIABLE_NAME));
-                }
-                catch (Exception)
-                {
-                    //no action
-                }
-                if (!supressWarningOrError)
-                {
-                    WriteWarning(Constants.MSGraphMigrationMessage);
-                }
+                }              
 
                 acsServicePrincipal = BuildServicePrincipal(Name, clientSecret);
                 WriteVerbose(Resources.CreatedANewServicePrincipalAndAssignedTheContributorRole);